// src/app/interfaces/divipola.ts
export interface Divipola {
    statename:[];
    distance: string;
    elevation: string;
    osmtags: {
        wikipedia: string;
        wikidata: string;
        source: string;
        isInState: string;
        divipola: string;
        name: string;
        clcfile: string;
        DANEDepartamento: string;
        adminLevel: string;
        objectid: string;
        boundary: string;
        DANEArea: string;
        isInCountry: string;
        type: string;
        entidadTe: string;
    };
}
// {
//    "statename":{ },
//    "distance":"0.104",
//    "elevation":"2546",
//    "osmtags":{
//       "wikipedia":"es:Funza",
//       "wikidata":"Q608955",
//       "source":"OCHA - SIGOT",
//       "is-in-state":"Cundinamarca",
//       "divipola":"25286",
//       "name":"Funza",
//       "clcfile":"MunicipiosDepartamentos/ADM_Municipios.shp",
//       "DANE-departamento":"25",
//       "admin-level":"6",
//       "objectid":"3847",
//       "boundary":"administrative",
//       "DANE-area":"71.00",
//       "is-in-country":"Colombia",
//       "type":"boundary",
//       "entidad-te":"Cabecera Municipal"
   
// },
//    "state":"CO",
//    "latt":"4.70456",
//    "city":"Funza",
//    "prov":"CO",
//    "geocode":"FACATATIVA-RIEBD",
//    "geonumber":"3670307534845",
//    "country":"Colombia",
//    "stnumber":{

   
// },
//    "staddress":"Carrera 2B",
//    "inlatt":"4.70378",
//    "alt":{

   
// },
//    "timezone":"America/Bogota",
//    "region":"Funza, Cundinamarca",
//    "postal":{

   
// },
//    "poi":{
//       "source":"Kaart Ground Survey 2017",
//       "poilat":"4.70589",
//       "name":"Farma Funza",
//       "addr-street":"Calle 15",
//       "amenity":"pharmacy",
//       "poilon":"-74.20461",
//       "id":"5173864524",
//       "poidist":"0.302",
//       "healthcare":"pharmacy"
   
// },
//    "longt":"-74.20684",
//    "remaining_credits":{

   
// },
//    "confidence":"0.9",
//    "inlongt":"-74.20633",
//    "class":"residential",
//    "adminareas":{
//       "admin6":{
//          "wikipedia":"es:Funza",
//          "source":"OCHA - SIGOT",
//          "divipola":"25286",
//          "clcfile":"MunicipiosDepartamentos/ADM_Municipios.shp",
//          "DANE-departamento":"25",
//          "admin-level":"6",
//          "boundary":"administrative",
//          "is-in-country":"Colombia",
//          "entidad-te":"Cabecera Municipal",
//          "wikidata":"Q608955",
//          "is-in-state":"Cundinamarca",
//          "name":"Funza",
//          "objectid":"3847",
//          "level":"6",
//          "DANE-area":"71.00",
//          "type":"boundary"
      
// }
   
// },
//    "altgeocode":"ELRANCHOCO-RIEBD"
// }